requirements.txt
streamlit
google-cloud-aiplatform==1.55.0
google-cloud-storage
google-auth
python-docx